requirements.txt
python-neutronclient
websocket-client==0.37.0
